TIP-841: Move the channel locale subscriber to Akeneo\Channel
